pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fairyclaw"
version = "0.1.1"
description = "FairyClaw server-side agent runtime"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "sqlalchemy>=2.0.30",
  "aiosqlite>=0.20.0",
  "pydantic-settings>=2.3.4",
  "httpx>=0.27.0",
  "websockets>=16.0",
  "aiofiles>=24.1.0",
  "python-multipart>=0.0.9",
  "pyyaml>=6.0.2",
  "ddgs",
  "beautifulsoup4>=4.12.0",
  "filetype>=1.2.0",
  "qdrant-client>=1.10.1",
  "tiktoken>=0.7.0",
]

[project.optional-dependencies]
# Local neural embeddings (pulls PyTorch + sentence-transformers; not required for hashing or API embeddings).
embedding-local = ["sentence-transformers>=3.0.1"]

[tool.setuptools]
packages = ["fairyclaw"]

[tool.uvicorn]
factory = false
host = "0.0.0.0"
port = 8000
